requirements.txt
asgiref==3.4.1
autopep8==1.6.0
dj-database-url==0.5.0
Django==3.2.8
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
psycopg2==2.9.2
toml==0.10.2
gunicorn==20.1.0
python-telegram-bot==12.7